refactor(web): split app.ts and components.ts into modules under 500 lines
The last two oversized files. Pure mechanical module extraction — functions
moved verbatim, imports fixed, no logic/template changes. tsc + eslint + build
all green.

components.ts (1913) -> a barrel re-exporting per-view component modules under
web/src/components/ (shared, issues-list, issues-detail, releases, alerts,
settings, logs, analytics). app.ts's import is unchanged.

app.ts (2741) -> a 5-line entry that imports ./bootstrap.js. Logic extracted
into core.ts (shared state/elements/DOM + util singletons), http.ts, live.ts,
router.ts, bootstrap.ts, and per-view modules (views-issues, views-issue-detail,
views-releases, views-alerts, views-logs, views-settings, views-projects). ES
module function hoisting handles the runtime-only router<->view<->http call
cycles; no module-eval-time init-order hazards.

Every tracked source file in the repo is now under the 500-line limit: the
check-file-length.sh allowlist is emptied. Largest file is now 492 lines.
